Enhance/custom status bar better caching (#60)
* Add SimVarLookup dataclass

This stores state for one lookup.  This will pave the way for allowing more dynamic lookup logic.
- functionality otherwise kept the same for now

* Add _is_aircraft_requests_defined helper function

* Remove continue so retry connection sleep logic works

* Update sim_connected == False case

- increase sleep
- add clarifying comment

* Fix _is_aircraft_requests_defined

* SimVarLookup - add value member

* vars_to_update - prioritize by last_update

* get_simconnect_value - fix logical error with checking value

* Add helper function: is_value_valid to get_simconnect_value

* Add reset_cache()

- if disconnect occurs this is called to reset cache to empty state

* SimVarLookup: add get_value

This will enforce by age to minimize risk of stale values

* Revise comments

* check_cache: fix issue with reference to function

get_value missing ()

* Add todo

* _is_aircraft_requests_defined - fix

* Tweak SimConnect aq values

Set to 10ms from 0 to get a tiny bit of SimConnects own caching, but set attempts to 2 so it will not retry default of 10 times (may be excessive for real time use like this).

* get_simconnect_value - add clarifying comment on purpose of retries

Since SimConnect has its own retry mechanism it may otherwise not be obvious why it exists

* get_simconnect_value - add further comments

* BackgroundUpdater - add slightly lower variable_sleep

10 ms was probably too high if many variables are present.  A little bit of sleep is required though to ensure the background updater doesn't restrict the main thread too much (since Python GIL restricts true multithreading behavior).

* Delete status_bar_templates.py

- old version incorrectly saved this here (Scripts/Settings rather than /Settings)

* simbrief_settings is no longer global

Based on the new refactored design, I don't feel that this belongs any longer as a global var.

This will break Leftos previous template, but only a couple lines on that will need to be changed.

* Rename check_cache to get_cache_value

* reset_cache - only show warning if there isn't a connection

* SimVarLookup: add set_value

* BackgroundUpdater: fix issue with run()

Finally able to test new changes on the sim to catch bugs with new implementation.
- sim_variables was not addressed properly with string key

* SimVarLookup: handle strings and validate type

SimConnect lib will either pass a float or string so now set_value will explicitly check for that.

Note that the SimConnect DLL represents strings as byte sequences, requiring explicit UTF-8 decoding.