Launcher - add script auto restart
If a script fails unexpectedly, it will restart the script after a small delay.
- The long term goal should be have this based on a per script meta-data setting.

A bit 'hackish' but since we seem to have a small chance of random failure on the init of the Mobiflight scripts this will allow these to restart automatically, which ultimately is better for the user.